+++ src/shared/common/hooks/useLocalStorage/types/UseLocalStorage.types.ts
