+++ src/shared/common/hooks/useSnackbar/types/index.ts
