+++ src/shared/common/hooks/useSnackbar/types/useSnackbar.types.ts
